Completed another challenge today!
Roman number to integer. This was a fairly simple task, by combining dictionary mapping with loop iterations. I basically just map roman literals to an integer number and access accordingly while looping and summing it up.

HOWEVER, there are edge cases to be considered such IV as it's a subtraction case due to I being in front of V hence the if loop.